Move Hello World Main to Folder Hello World
Move the Hello World project to the folder Hello World to make way for
the next project, as they will be multiple projects in this repo since
each project is too small to be on its own. However, they may get their
repo if those projects are big enough, but for now, to keep it simple
will be moving the project around with the active one in the root
directory.