Move Dapper Dasher Main to Folder Dapper Dasher
Move the Dapper Dasher project to the folder Dapper Dasher to make way
for the next project, as they will be multiple projects in this repo
since each project is too small to be on its own. However, they may get
their repo if those projects are big enough, but for now, to keep it
simple will be moving the project around with the active one in the root
directory.